organize a little better
