+++ pkg/api/parts.go
@@ -3,26 +3,10 @@ package api
 import (
 	"bytes"
 	"context"
-	"fmt"
-	"google.golang.org/api/sheets/v4"
 	"net/http"
 	"path/filepath"
 	"strings"
 )
-
-type Part struct {
-	Project            string
-	ProjectTitle       string `col_name:"Project Title"`
-	PartName           string `col_name:"Part Name"`
-	ScoreOrder         int    `col_name:"Score Order"`
-	SheetMusicFile     string `col_name:"Sheet Music File"`
-	ClickTrackFile     string `col_name:"Click Track File"`
-	ConductorVideo     string `col_name:"Conductor Video"`
-	Released           bool
-	Archived           bool
-	ReferenceTrack     string `col_name:"Reference Track"`
-	PronunciationGuide string `col_name:"Pronunciation Guide"`
-}
 
 type PartView struct {
 	SpreadSheetID string
@@ -61,33 +45,6 @@ func (x PartView) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	renderPartsView(w, ctx, wantProjects, parts, x.Distro.Name)
 }
 
-func listParts(ctx context.Context, spreadSheetID string) ([]Part, error) {
-	srv, err := sheets.NewService(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to retrieve Sheets client: %w", err)
-	}
-
-	readRange := "Parts"
-	resp, err := srv.Spreadsheets.Values.Get(spreadSheetID, readRange).Do()
-	if err != nil {
-		return nil, fmt.Errorf("failed to retrieve data from sheet: %w", err)
-	}
-
-	if len(resp.Values) < 1 {
-		return nil, fmt.Errorf("no data")
-	}
-	parts := make([]Part, len(resp.Values)-1)
-
-	index := make(map[string]int, len(resp.Values[0])-1)
-	for i, col := range resp.Values[0] {
-		index[fmt.Sprintf("%s", col)] = i
-	}
-
-	for i, row := range resp.Values[1:] {
-		processRow(row, &parts[i], index)
-	}
-	return parts, nil
-}
 
 func renderPartsView(w http.ResponseWriter, ctx context.Context, projects []Project, parts []Part, distroBucket string) {
 	type tableRow struct {
@@ -135,10 +92,3 @@ func renderPartsView(w http.ResponseWriter, ctx context.Context, projects []Proj
 	_, _ = buffer.WriteTo(w)
 }
 
-func downloadLink(bucket, object string) string {
-	if bucket == "" || object == "" {
-		return ""
-	} else {
-		return fmt.Sprintf("/download?bucket=%s&object=%s", bucket, object)
-	}
-}
